Put tests inside main package so other packages can access
